requirements.txt
Flask
flask-cors
pdfplumber
python-dotenv
requests
groq
python-pptx